requirements.txt
django>=4.2
djangorestframework>=3.14
channels>=4.0
django-cors-headers>=3.13
psycopg2-binary>=2.9
daphne>=4.0
whitenoise>=6.2
django-environ>=0.9
openai>=0.27  # for LLM integration
redis>=4.5  # for channels layer (optional but recommended)
